MDL-80111 h5p: Validate Moodle language code with H5P language codes.
After getting the current Moodle language code, validate it with the
available H5P language codes. If the code is not valid, remove the last
part of the code and validate it again in a loop.
If the string is "pt_br_wp" it will return "pt_br", and if the string is
"en_us_wp" it will return "en", because "en_us" is not a valid H5P
language code.